+++ packages/generator/src/generator.ts
@@ -26,11 +26,7 @@ import { logger } from './lib/logger'
 import { getEnumModuleName } from './lib/prisma-helpers/enums'
 import { isRelationField } from './lib/prisma-helpers/field'
 import { ImportValue, namedImport, NamedImport } from './lib/syntaxes/imports'
-import {
-  createModelModule,
-  ModelModule,
-} from './lib/syntaxes/modules/model-module'
-import { createModule, Module } from './lib/syntaxes/modules/module'
+import { createModule, Module } from './lib/syntaxes/module'
 
 const { version } = require('../package.json')
