+++ packages/generator/src/lib/adapter/adapter.ts
@@ -1,6 +1,6 @@
 import { PrismaEnumField, PrismaScalarField } from '../prisma-helpers/field'
 import { ImportValue } from '../syntaxes/imports'
-import { Module } from '../syntaxes/modules/module'
+import { Module } from '../syntaxes/module'
 import { FieldFunc } from './fields/createField'
 
 type ParsableField = PrismaScalarField | PrismaEnumField
