+++ packages/generator/src/lib/adapter/declarations/generateTableRelationsDeclaration.ts
@@ -10,7 +10,6 @@ import {
 } from '~/lib/prisma-helpers/field'
 import { getDbName } from '~/lib/prisma-helpers/getDbName'
 import { getModelVarName } from '~/lib/prisma-helpers/model'
-import { ModelModule } from '~/lib/syntaxes/modules/model-module'
 import { namedImport } from '../../syntaxes/imports'
 
 type GenerateTableRelationsInput = {
